refactor - modify menu groups package structure
- 메뉴 그룹과 관련된 패키지 구조를 애그리거트 중심 구조로 변경했습니다
