+++ src/main/java/kitchenpos/menugroups/dao/JdbcTemplateMenuGroupDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.menugroups.dao;
 
-import kitchenpos.model.MenuGroup;
+import kitchenpos.menugroups.model.MenuGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
