+++ src/test/java/kitchenpos/controller/MenuGroupRestControllerTest.java
@@ -1,7 +1,8 @@
 package kitchenpos.controller;
 
-import kitchenpos.bo.MenuGroupBo;
-import kitchenpos.model.MenuGroup;
+import kitchenpos.menugroups.bo.MenuGroupBo;
+import kitchenpos.menugroups.controller.MenuGroupRestController;
+import kitchenpos.menugroups.model.MenuGroup;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
